backend/requirements.txt
# Django y REST Framework
Django==3.2.24
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-ratelimit==4.1.0
drf-spectacular==0.28.0

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# Caché y colas asíncronas
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.0
kombu==5.3.4

# Almacenamiento
boto3==1.34.14
django-storages==1.14.2

# Machine Learning
numpy==1.26.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
xgboost==3.0.4
matplotlib==3.10.5

# Utilidades
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
psutil==5.9.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# PDF generation
markdown2pdf==0.1.0
reportlab==4.1.0

